pyproject.toml
[project]
name = "challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.0",
    "httpx==0.25.0",
    "langchain>=0.0.267",
    "pydantic>=2.0.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests>=2.32.3",
    "smolagents[litellm]>=1.13.0",
    "uvicorn>=0.34.0",
    "websockets==11.0.3",
]
